moved changes to separate component
