+++ ui/pages/sigmaker.tsx
@@ -3,7 +3,6 @@ import React, { useState } from 'react'
 import {
   Divider,
   Grid,
-  Link,
   ThemeProvider,
   Typography
 } from '@mui/material'
@@ -20,32 +19,18 @@ import { getServerSideSessionOrRedirect } from '../server/getServerSideSessionOr
 import EmailSignature from '../components/emailSignature/emailSignature'
 import { icons } from '../styles/icons'
 import EmailSignatureForm from '../components/emailSignature/emailSignatureForm'
-import ArrowBackIcon from '@mui/icons-material/ArrowBack'
+import BackArrow from '../components/backArrow/backArrow'
 
 const Sigmaker: NextPage = () => {
   const [signatureData, setSignatureData] = useState<undefined | SignatureData>(
     undefined
   )
 
-  const [isHover, setIsHover] = useState(false)
-
-  const backBtnStyle = {
-    color: isHover ? 'darkblue' : 'green'
-  }
-
   return (
     <Layout>
       <ThemeProvider theme={theme}>
         <StyledPageContainer>
-          <Link href="/">
-            <span
-              style={backBtnStyle}
-              onMouseEnter={() => setIsHover(true)}
-              onMouseLeave={() => setIsHover(false)}
-            >
-              <ArrowBackIcon fontSize="large" />
-            </span>
-          </Link>
+          <BackArrow />
           <Typography variant="h3"> Signature Maker </Typography>
           <Divider />
           <br />
